requirements.txt
openai>=1.57.0
chromadb>=0.5.11
langchain-openai>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.4
faiss-cpu>=1.8.0
open_clip_torch>=2.24.0
torch>=2.2.0
pillow>=10.4.0
requests>=2.32.0
python-dotenv>=1.0.1
langchain-experimental>=0.0.67
rank-bm25>=0.2.2
FlagEmbedding>=1.2.10

